Switch to webpack
I realize that rollup is going to be a little annoying to work with in
the test environment. Its main use-case is to provide a bundle free from
outside dependencies, one that can be easily packaged as an npm module.
I have a different use-case: I want to provide a single bundle to try
out the component locally.

I might bring back rollup for the exported module later.