Snap to nearest 5-min mark
We want to allow a little better control when selecting times. Snapping
to closest 5-min mark will make it easier to target when to start and
stop.

Also updates each hour to be 60px tall. Makes calculations a lot
simpler. We'll see how this fans out since it makes the calendar a
little too tall imo.