Make overlapping work somewhat okay
I took a different approach to overlapping. First, events are grouped
and put in columns. Then, we flatten the groups, adding width and offset
information to the event. This information is then used in <TimeSlot> to
make overlapping events render next to each other.

One enhancement I want to make is to let certain events span multiple
columns based if the neighboring columns aren't filled at the current
slot.