Sync header width with days wrapper width
This fixes an issue with how a scrollbar can shift the days content
~15px to the left. We had previously fixed that by adding a default 15px
padding-right to the header (0c29873), but that made it look odd in
browsers that render the scrollbar semi-transparent on top of the
content (hence not eating up any width for the content).

We can fix this by keeping the two element widths in sync.