Fix incorrect times when recurring and DST week
Due to how we reuse a week for the recurring view, we risk creating
recurring times on a DST week. In such cases, the
minutes-since-start-of-week calculation was wrong. A 4:00 - 5:00
selection would be represented as 180 - 240 minutes.

By moving to the first week of the year, we should avoid most DST
issues. I found this resource that lists time changes, and it doesn't
look like any country switches on the first week:

https://www.timeanddate.com/news/time/

(There are changes to timezones and such, but no recurring DST switches)