Make <Day> work for devices with mouse AND touch
Our feature-detection was cancelling out mouse events from <Day> if
touch was detected. This meant that a device with both touch and mouse
events wouldn't be able to select time slots.

Fixed by always attaching mouse listeners, even on touch devices. The
only difference we found was that the ios simulator in xcode showed a
short "click" flash after adding a slot. But things seemed to work.